refactor(auth): move google auth mutation to hooks directory
Move the Google authentication mutation hook from the api directory to the features/user/hooks/auth directory for better organization and maintainability
